refactor: RacingCarContorller 및 내부 method 이름 변경 / service의 makeCars, makeCount 로직 위치 controller로 변경
